Moving RetryHelper tests to a new file.
